Improve docker deployment script safety, logging clarity and cleaning up

+++ deployments/deploy-types.sh
@@ -21,6 +21,11 @@ function deploy_default() {
     echo -e "Starting docker build\n"
     docker build -f Dockerfile-prod -t "$image_tag" .
     local build_status="$?"
+
+    # Disabled as this meant that no dependencies could be cached. Instead run `docker system prune` periodically to clear up disk space if necessary.
+    # echo "Removing intermediate **images** left by docker build (when using multistage dockerfiles)"
+    # docker image prune -f --filter label=stage=builder
+
     if [ "$build_status" != 0 ]; then
 	>&2 echo -e "\n###-> ERROR! Build failed! Aborting deployment!"
 	return "$build_status"
@@ -28,25 +33,53 @@ function deploy_default() {
 
     # Stopping old docker container and waiting for it to exit
     echo -e "\n###-> Build done. Stopping old container and waiting for it to exit\n"
-    docker stop "$old_container_id"
-    docker wait "$old_container_id"
-    
-    echo -e "\n###-> Old container stopped, running new one\n"
+    if [[ -n "$old_container_id" ]]; then
+	docker stop "$old_container_id" &>/dev/null
+	printf "old container exit code: "
+	docker wait "$old_container_id"
+	echo -e "\n###-> Old container stopped.\n"
+    else
+	echo "###-> No container was previously running! Continuing..."
+    fi    
+
+    echo -e "###-> Running new container\n"
     docker run -d --restart=unless-stopped "$image_tag"
     local run_status="$?"
     if [ "$run_status" != 0 ]; then
 	>&2 echo -e "\n###-> ERROR! Run failed!"
 	>&2 echo "###-> Retagging old image and starting old container back up"
-	docker tag "$old_image_id" "$image_tag"
-	docker start "$old_container_id"
+	if [[ -n "$old_image_id" ]]; then
+	    docker tag "$old_image_id" "$image_tag"
+	else
+	    echo "###->> No old image found for retagging!!"
+	fi
+	if [[ -n "$old_container_id" ]]; then
+	    docker start "$old_container_id"
+	else
+	    echo "###->> No old container found for starting back up!!"
+	fi
 	return "$run_status"
     fi
 
     # Cleanup
     echo -e "\n###-> New container now running successfuly, removing old container and image!"
-    docker rm "$old_container_id"
-    docker rmi "$old_image_id"
-    
+    if [[ -n "$old_container_id" ]]; then
+	printf "old container id: "
+	docker rm "$old_container_id"
+    else
+	echo "###-> No old container was running, so none removed."
+    fi
+    if [[ -n "$old_image_id" ]]; then
+	if [[ "$(docker images -q "$image_tag")" == "$old_image_id" ]]; then
+	    echo "###-> Not removing image, as the container was run using the same one (build did a full cache hit)"
+	else
+	    printf "old image id: "
+            docker rmi "$old_image_id"
+	fi
+    else
+	echo "###-> No old image found, so none removed."
+    fi
+
     return "$run_status"
 }
